Updated Layer 2 code with RS <-> DS sequence tracking and error correction (#101)
* Refactor tcs protocol handle info into smaller functions.

* New module to handle TCS packet windowing structs and functions.
Checks on making sure packets are received in order and inside the current receive window.

* Refactoring and documentation, prep for adding windowing checks.

* Added debugging statements.
Counts packets used to know how much to advance the window start.

* Changes to debugging statements.

* Last changes that use TCS window for monitoring only.

* Changed references of "window" to "rx_window" in case there is a separate tx one later.

* Documenting functions.
Adding simple utility functions.
Trimming down module names in specs.

* Starting to use sequences from the tcs Window struct instead of rx_seq, for packets
outside the current window and out of sequence packets.

* Incoming dropped packet handling.
Structure and error checking code for ensuring
incoming packets are in the proper order.
The code checks that there are no pending packets before the received packets, and if there are
out of order packets it returns the action to do, which will be translated into a function call.

* Added logging statements to help debugging.

* Code to handle missing packets from the RS.
Formatting changes.
Adding packet numbers to logging of acks and wacks.

* Added sequence numbers to debugging statements.

* Added cachex for holding transmitted packets
until acked.

* Set up cache for transmitting packets.

* Added more information to debugging logs.
Add packets to cache on original transmit, delete on ack, resend on nakcce.

* Move the sending of packets from the main ranch handle_info loop to
its own GenServer, Transmitter. It checks to see if the RS is ready to accept packets before sending them.

* Changed Transmitter module to its full name, Prodigy.Server.Protocol.Tcs.Transmitter

* Sometimes RS sends a nakcce then a nakccc for a particular packet.
Transmitter now keeps track of when a nakcce is replied to, and doesn't reply
to a nakcce on the same packet.

* Added the ack_tracker cache that will keep up with
if acks have already been sent to RS.

* Added the ack_tracker cache that will keep up with
if acks have already been sent to RS.
If DS gets a wackpk on a packet sequence but DS already sent an ackpkt, DS just sends another ackpkt.
Fixed the internal typo, nakccc -> nakncc.

* Adjusted expiry times to prepare for upcoming wackpk sending.

* Added fields to the tuple to keep track of the number of wackpk's sent, plus the time
of the last one sent so DS knows when to send one.
Added the check in the main loop to see if it's time to send any wackpk's.

* Send self pid to Transmitter so that it knows where to send the wackpk limit exceeded message.
Handle the wackpk limit exceeded (:wp_limit_exceeded) message to know to close this connection.

* A classic off-by-one error.
Set so that wackpk attempts @wack_threshold (4) times correctly.

* Change the items stored in the cache from a tuple to the struct PacketState.
The tuple was getting too many members to keep track of.

* Started keeping acked packets around, just marking acked as true.
Using rs_buf to keep the list of "starting unacked" list of packets in case we get a rxmitp.
After each ack, recalculate the rs_buf, trimming acked packets from the head of the list.
If we get a rxmitp, send all the packets in rs_buf.

* Handle rxmitp, send packets from DS.

* Changed logging severity on packet outside window.

* Added extra conditionals so that if we receive a packet outside the receive window
then we send a rxmitp and we don't move up the receive window.

* Updated tests to math new code to handle
transmission errors in both directions.
Currently no errors in tests.

---------

Co-authored-by: Phillip Heller <pheller@me.com>